Remedy for when indexer is down.
Keep track of the block number when last event is caught (blockOfLastEvent), and start syncing the indexer from this block. This will let indexer catch the events when it was down.

Introduces a collection BlockChainReviewMeta to keep blockOfLastEvent. Ideally keep a document with _id: 'meta' in BlockChainReview but mongoose does not allow this as it does not fit the schema. This is an ok workaround.