Don't show vouch button for own reviews at VouchReview
Also changes how author names are displayed on VouchReview table. Previously the author address was replaced in review array, on data level. Now the address is kept in the array and only name is displayed when rendering. No changes to data.

It does not make sense to be able to vouch for own reviews. This is implemented on contract level. Changes reflect the implementation on UI.