Stop server-side paginating indexed reviews, add comments
We used to use mongoose-paginate to send indexed reviews in a paginated manner for example: 35 reviews by batches of 10 in 4 pages.

We wanted to search the reviews in VouchReviews both by BlockchainReview fields and by Author name and email. This turns out to be difficult. We need to do a text .find() on several fields of BlockchainReviews and return those. Plus do a text .find() on author fields name and email, do a populate() for the reviews of matching authors and append those reviews to the result. In the end the search results will come from two sources BlockchainReviews and ReviewAuthors and the results will include duplicate results. Also we can't easily paginate those combined results.

Therefore we give pagination and filtering server side up and send all reviews to the client. Filtering shall be done on client side.