Add search through GlobalFilter to VouchReview
Now that we have all indexed reviews in the table we utilze the
GlobalFilter of the react-table for searching reviews.

Since the received review objects only had author addresses we were
populating author names externally by fetching an 'authorsMap' from the server, check the
address as key and display names on the table. The data in review object
remained same as the address. Therefore we were not able to search
author names on the table.

Instead now we receive review objects with author info populated server
side. We define new cells First Name and Last Name when declaring the
table and we are then able to search the names when we perform a global
search. By default this allows searching other fields that are in text
or number format such as publisher, timestamp... However date fields
require additional formatting when searching.