Fix indexer registering multiple Event listeners
When connected to bloxberg, the indexer seems to start a new connection
to bloxberg WebSocketProvider. Since there there is an event listener
.on('connect') to the WebSocket which initializes the indexer app, this
initializer method is called each time the indexer reconnects to
WebSockets provider which in turn adds an Ethereum Event listener for
ReviewAdded, ReviewVouched, and ReviewDeleted. As a result when a Review
is added, if the indexer reconnects 10 times, the review is added to the indexer database in BlockchainReview model.

The commit introduces a workaround by keeping a variable isInit, and
registers Ethereum Events for ReviewAdded, ReviewVouched, ReviewDeleted
only once.