Promt user to log in w/ Metamask instead of auto
Make users click to "Log in with Metamask" button in the welcome login
screen to log in with Metamask. Previously we were checking if Metamask
is injected and were calling ethereum.enable() automatically, which is
not the recommended way.

Also store the web3 object in App.js state and in
Context to possibly use in other modules. Ideally other modules should
call this Web3 object instead of creating a new one every time. (Which I
suspect causing bad user experience currently)

Introduce minor styling changes in the LoginPage.