Fix Login page loading too slow
Change the ordering of how components loaded to quickly
respond to the user.

Also add web3 and instance objects to App.js state and
serve web3 in Context. But couldn't make use of this as
I couldn't figure out how to reuse these in reviewConnection.js.

Does reviewConnection.js also have to be a React component?
How can we create the web3 object and the conttract instance
object once, store it somewhere (App.js state?) and use it every
time we make a call to the blockchain. Currently we create web3
and the contract instance upon every call in reviewConnection.js
init() function.