Add settings page and update user on submit
Create a settings page where the user can change their name and mail
information. Refactor backend code and get rid of /register/ endpoints.
Move the funcitons to /authors/ endpoint. Add update function for user.

Add props.children to FormField, to add buttons to the right of the
Input field.

However, one problem noticed happens when both Chrome (with Metamask)
and Firefox (with Fortmatic) are open and the user submits on Firefox.
Interestingly from Firefox, the name and mail of the Firefox user with
the address of the _Chrome user_ is pushed. Must be inspected deeper.
Pushing to open a new branch.