Fix adding too many event listeners at setProvider
When calling setProvider() on the TruffleContract object, it seems it is
registering an event listener. Since we were calling this every time
when we make a contract call, this registers too many event listeners.
This can be one of the reasons why the app sometimes feels really slow
and even makes some PCs stuck when in the dev environment.

This was achieved by making the getWeb3() function at /web3.js
synchronous and by moving the setProvider() function outside the async
init() function in reviewConnection.js. Ideally we also move the
contract instance after .deployed() but I couldn't figure that out yet.
Current status should be sufficient and accellerate the app.